requirements.txt
bleak
pycryptodome
requests

